app/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
frozenlist==1.4.0
greenlet==3.0.2
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
yarl==1.9.4
